Fixed picture upload bugs.
This version of picture upload is made as simple as needed for it to
work. For now you can only upload when editing the listing, and it does
not show already uploaded files or allow you to delete them.
Known Issue: Should it fail to save the file there may be entries in the
database where fileName and position is NULL.